refactor : guide page component, hooks, lib 분리
